+++ src/models/spar_models.py
@@ -216,7 +216,7 @@ class SparSales(BaseModel):
     product: str
     quantity: float
     receipt_row_number: int | None = None
-    value: int | None = None
+    value: float | None = None
     sales_value_with_tax: float | None = None
     tax_amount: float | None = None
     sales_tax_rate: float | None = None
@@ -224,6 +224,7 @@ class SparSales(BaseModel):
     #campaign_code: str | None = None
     #transaction_campaign_code: str | None = None
     custom_supplier_cost: float | None = None
+    #loyalty_customer_type: str | None = None
 
 class SparPromotionsCompetitorPrices(BaseModel):
     observed_at: datetime
